feat(chat): 컴포넌트 이름 변경 및 채팅방 id 전달
수정 사항

- ChatRoomList -> ChatroomItem 컴포넌트 이름 변경
- 채팅방 id를 넘겨줌 -> navigate에서 id 값을 채팅방에 전달해주기 위해
